moving resource form to pages (it's a page)
