+++ client/src/app/pages/resource-form/resource-form.component.scss
